complete reorg of solar and thermal in envphys to make it easier to understand where functions will be
