+++ smrf/envphys/constants.py
@@ -1,0 +1,19 @@
+# define some constants
+MAXV = 1.0              # vis albedo when gsize = 0
+MAXIR = 0.85447         # IR albedo when gsize = 0
+IRFAC = -0.02123        # IR decay factor
+VFAC = 500.0            # visible decay factor
+VZRG = 1.375e-3         # vis zenith increase range factor
+IRZRG = 2.0e-3          # ir zenith increase range factor
+IRZ0 = 0.1              # ir zenith increase range, gsize=0
+STEF_BOLTZ = 5.6697e-8  # stephman boltzman constant
+EMISS_TERRAIN = 0.98    # emissivity of the terrain
+EMISS_VEG = 0.96        # emissivity of the vegitation
+FREEZE = 273.16         # freezing temp K
+BOIL = 373.15           # boiling temperature K
+STD_LAPSE_M = -0.0065   # lapse rate (K/m)
+STD_LAPSE = -6.5        # lapse rate (K/km)
+SEA_LEVEL = 1.013246e5  # sea level pressure
+RGAS = 8.31432e3        # gas constant (J / kmole / deg)
+GRAVITY = 9.80665       # gravity (m/s^2)
+MOL_AIR = 28.9644       # molecular weight of air (kg / kmole)
