+++ smrf/envphys/solar/ipw.py
@@ -2,6 +2,7 @@ import os
 import subprocess as sp
 
 import numpy as np
+
 on_rtd = os.environ.get('READTHEDOCS') == 'True'
 if on_rtd:
     IPW = '.'  # placehold while building the docs
